requirements.txt
tensorflow>=2.19.0
pillow>=11.2.0
numpy>=2.1.3
